+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+### Feature
+- Add in Address Contract Details [#914](https://github.com/EthVM/EthVM/pull/914)
 
 ### Release v1.0.3-hotfix.1