+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 - create workflow for master and staging [#915](https://github.com/EthVM/EthVM/pull/915)
 
+### Release v1.0.3-hotfix.2
+
+- Fix variables [#908](https://github.com/EthVM/EthVM/pull/908)
+
 ### Release v1.0.3-hotfix.1
 
 ### Devop